create table

+++ controler.py
@@ -1,3 +1,5 @@
+from os import write
+
 from model import Model
 from view import View
 from PyQt5 import QtCore, QtGui
@@ -18,6 +20,7 @@ class Controller:
         self.t_red = []
         self.t_yellow = []
         self.t_green = []
+        self.dataTabel = None
         #
         self.setSylesheet()
 
@@ -42,7 +45,7 @@ class Controller:
 
         src = QFileDialog.getOpenFileName(filter="Video (*.*)")[0]
         self.vid = cv2.VideoCapture(src)
-        # src = './dataset/dataset-5s.mp4'
+        # src = './dataset/dataset-2s.mp4'
         # self.vid = cv2.VideoCapture(src)
 
         if not self.vid.isOpened():
@@ -59,6 +62,7 @@ class Controller:
             self.clear()
 
         self.playing = True
+        self.view.grafView.clear()
 
         # memastikan agar list tersebut kosong
         self.t_red = []
@@ -99,6 +103,8 @@ class Controller:
     def update_frame(self):
         import timeit
 
+        ''' data yg akan dimasukkan kedalam tabel frame, persentase hijau, kuning, merah '''
+
         if self.playing:
             ret, frame_ori, hsv, hsv_vid_out, red, yellow, green = self.model.color_map(self.vid, self.view.radioButton_record.isChecked(), self.t_px)
             if ret:
@@ -118,6 +124,9 @@ class Controller:
                 self.vid.release()
                 if self.view.radioButton_record.isChecked():
                     self.video_out.release()
+
+                self.table()
+
                 self.vid = None
                 self.view.label_status.setText("Done")
                 # labels = [self.view.label_dp, self.view.label_ori, self.view.label_indi_dp]
@@ -148,11 +157,18 @@ class Controller:
     def save(self):
         import shutil, os
 
-        if os.path.exists('./result/result.mp4'):
-            os.remove('./result/result.mp4')
-        shutil.move('result.mp4', './result')
+        if self.dataTabel != None:
+            if self.view.radioButton_record.isChecked():
+                if os.path.exists('./result/result.mp4'):
+                    os.remove('./result/result.mp4')
+                shutil.move('result.mp4', './result')
 
-        self.view.label_status.setText("Saved to directory result")
+            self.dataTabel = self.dataTabel.replace("\t", ",")
+            with open('./result/dataTable.csv', 'w') as file:
+                file.write(self.dataTabel)
+            self.dataTabel = None
+
+            self.view.label_status.setText("Saved to directory result")
 
 
     def clear(self):
@@ -164,6 +180,8 @@ class Controller:
         [label.setText(" ") for label in labels]
 
         self.view.grafView.clear()
+
+        self.view.tableView.clear()
 
     def change_icon_play(self, condision):
         src = "icon/play.png" if condision == 1 else "icon/pause.png"
@@ -183,6 +201,32 @@ class Controller:
         self.line_red.setData(x_r, self.t_red)
         self.line_yellow.setData(x_y, self.t_yellow)
         self.line_green.setData(x_g, self.t_green)
+
+    def table(self):
+        hd = ['Frame', 'Green', 'Yellow', 'Red']
+
+        data = []
+
+        [data.append({hd[0]: i, hd[1]: f"{self.t_green[i]}%", hd[2]: f"{self.t_yellow[i]}%", hd[3]: f"{self.t_red[i]}%"}) for i in range(1, len(self.t_red))]
+
+        # for i in range(len(self.t_red)):
+        #     data.append({hd[0]: i, hd[1]: f"{self.t_green[i]}%", hd[2]: f"{self.t_yellow[i]}%", hd[3]: f"{self.t_red[i]}%"})
+
+        self.view.tableView.setData(data)
+        self.dataTabel = self.view.tableView.serialize(useSelection=False)
+
+        self.view.tableView.verticalHeader().setVisible(False)
+
+        self.view.tableView.setColumnWidth(0, 73)
+        self.view.tableView.setColumnWidth(1, 100)
+        self.view.tableView.setColumnWidth(2, 100)
+        self.view.tableView.setColumnWidth(3, 100)
+
+        for row in range(self.view.tableView.rowCount()):
+            for col in range(self.view.tableView.columnCount()):
+                item = self.view.tableView.item(row, col)
+                if item:
+                    item.setTextAlignment(QtCore.Qt.AlignCenter)
 
 
 def main():
